Capture a class instance instead of the member
In our last commit, we captured `ps.alphabeticPattern` because it was the most straightforward thing to do. However, if by some reason we need to access several members on a class (use case not depicted in the diff), we may want to capture the serializable instance instead of member by member.

This commit captures `ps`, object of class `Patterns`. By doing this, remember that the transitive checker will analyse Patterns. That's why we modified the class definition with `sealed`, to make sure that no subclass can ever be non-serializable.