Use `capture` instead of explicit header declaration
Spores allow you to define your environment in the same spore body. This is useful because you can avoid the "boilerplate" of defining a header with all the concrete environment (which is sometimes recommended). `capture` may ease the translation between anonymous functions and spores.

However, the use of `capture` is only limited to identifier. That means that we cannot do `capture(ps.alphabeticPattern)` but we have to explicitly capture the identifier `ps` with `capture(ps)`. This limitation makes sure that the body does not have any `this` reference.

The following commit transforms the more verbose spore definition into a simpler spore definition that uses `capture`.